organise RevTranslate code into separate file
